update "github" consent to retrieve the handle
The GitHub consent in AMY only gives consent for the github handle, not
the name associated with the handle:

> Yes, and only use my GitHub Handle (agree)

The previous iteration checked the GitHub API to get the name associated
with the handle. Given the wording of the AMY declaration, we can assume
that the handle is a better choice.