add catch in lesson creators to append mailmap
Since github emails contain the username, we can assume that we can
match their names in our database. If we can do that, then we can
construct a mailmap entry that will account for those.

This will search for those github email addresses, match to AMY, and
append the ones we found to the master mailmap.

It will then re-run the git log counter and give the results.